feat(resp): 实现 RESP 协议解析器和解码功能
- 添加了 Resp 接口的静态 decode 方法用于解析 RESP 协议数据
- 实现了对各种 RESP 类型的支持（状态、错误、整数、批量字符串、数组）
- 修改 RespInt 和 getNumber 方法中的 long 类型为 int 类型
- 添加了完整的 RESP 协议解析测试用例
- 将 STRING 类型重命名为 STATUS 类型以更准确反映协议规范

+++ src/main/java/cn/twopair/resp/RespInt.java
@@ -12,5 +12,5 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 public class RespInt implements Resp {
-	private final long value;
+	private final int value;
 }
